+++ data/protein.py
@@ -161,8 +161,8 @@ def to_pdb(prot: Protein, model=1, add_end=True) -> str:
   atom_mask = prot.atom_mask
   aatype = prot.aatype
   atom_positions = prot.atom_positions
-  residue_index = prot.residue_index.astype(np.int32)
-  chain_index = prot.chain_index.astype(np.int32)
+  residue_index = prot.residue_index.astype(int)
+  chain_index = prot.chain_index.astype(int)
   b_factors = prot.b_factors
 
   if np.any(aatype > residue_constants.restype_num):
